Organize, simplify.
