+++ project/README.md
@@ -8,5 +8,15 @@ apt-get install sox ffmpeg libcairo2 libcairo2-dev
 **In Linux terminal, install dependencies** 
 pip install -r requirements.txt 
 
-**Resources Used**
+**Resources and References**
 http://stackoverflow.com/questions/13728392/moving-average-or-running-mean
+
+----------------------------parseSong.py-----------------------------------------------------------------------------
+pip install soundfile
+all you need to install! Make sure to change directories in the main. 
+I am going to send the wav file inside the folder in the Github, place it in whatever directory you want :)
+REFERENCES
+https://docs.python.org/3/library/wave.html#module-wave
+https://stackoverflow.com/questions/69291258/reading-wav-as-bytes
+https://stackoverflow.com/questions/45010682/how-can-i-convert-bytes-object-to-decimal-or-binary-representation-in-python
+-----------------------------------------------------------------------------------------------------------------------
