+++ project/requirements.txt
@@ -1,6 +1,6 @@
 adafruit-blinka
 numpy
-playsound
+playsound==1.2.2
 pygobject
 pycario
 RPi.GPIO
