+++ project/tempPy.py
@@ -1,0 +1,6 @@
+from os import listdir
+from os.path import isfile, join
+
+mypath = "./audio"
+onlyfiles = [f for f in listdir(mypath) if isfile(join(mypath, f))]
+print(onlyfiles)
